Split out wrapper component from the page
